+++ Sources/SpeechlyAPI/speechly_config_v1_model_api.pb.swift
@@ -109,6 +109,8 @@ public struct Speechly_Config_V1_ListBaseModelsRequest {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
+  public var projectID: String = String()
+
   public var language: String = String()
 
   public var unknownFields = SwiftProtobuf.UnknownStorage()
@@ -230,7 +232,8 @@ extension Speechly_Config_V1_DownloadModelResponse: SwiftProtobuf.Message, Swift
 extension Speechly_Config_V1_ListBaseModelsRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementationBase, SwiftProtobuf._ProtoNameProviding {
   public static let protoMessageName: String = _protobuf_package + ".ListBaseModelsRequest"
   public static let _protobuf_nameMap: SwiftProtobuf._NameMap = [
-    1: .same(proto: "language"),
+    1: .standard(proto: "project_id"),
+    2: .same(proto: "language"),
   ]
 
   public mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
@@ -239,20 +242,25 @@ extension Speechly_Config_V1_ListBaseModelsRequest: SwiftProtobuf.Message, Swift
       // allocates stack space for every case branch when no optimizations are
       // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.language) }()
+      case 1: try { try decoder.decodeSingularStringField(value: &self.projectID) }()
+      case 2: try { try decoder.decodeSingularStringField(value: &self.language) }()
       default: break
       }
     }
   }
 
   public func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
+    if !self.projectID.isEmpty {
+      try visitor.visitSingularStringField(value: self.projectID, fieldNumber: 1)
+    }
     if !self.language.isEmpty {
-      try visitor.visitSingularStringField(value: self.language, fieldNumber: 1)
+      try visitor.visitSingularStringField(value: self.language, fieldNumber: 2)
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   public static func ==(lhs: Speechly_Config_V1_ListBaseModelsRequest, rhs: Speechly_Config_V1_ListBaseModelsRequest) -> Bool {
+    if lhs.projectID != rhs.projectID {return false}
     if lhs.language != rhs.language {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true