+++ cpp/Speechly/speechly/config/v1/model.pb.cc
@@ -23,7 +23,6 @@ constexpr BaseModel::BaseModel(
   ::PROTOBUF_NAMESPACE_ID::internal::ConstantInitialized)
   : name_(&::PROTOBUF_NAMESPACE_ID::internal::fixed_address_empty_string)
   , alias_(&::PROTOBUF_NAMESPACE_ID::internal::fixed_address_empty_string)
-  , is_available_(false)
   , is_downloadable_(false){}
 struct BaseModelDefaultTypeInternal {
   constexpr BaseModelDefaultTypeInternal()
@@ -49,7 +48,6 @@ const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_speechly_2fconfig_2fv1_2fmodel
   ~0u,  // no _weak_field_map_
   PROTOBUF_FIELD_OFFSET(::speechly::config::v1::BaseModel, name_),
   PROTOBUF_FIELD_OFFSET(::speechly::config::v1::BaseModel, alias_),
-  PROTOBUF_FIELD_OFFSET(::speechly::config::v1::BaseModel, is_available_),
   PROTOBUF_FIELD_OFFSET(::speechly::config::v1::BaseModel, is_downloadable_),
 };
 static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
@@ -62,16 +60,16 @@ static ::PROTOBUF_NAMESPACE_ID::Message const * const file_default_instances[] =
 
 const char descriptor_table_protodef_speechly_2fconfig_2fv1_2fmodel_2eproto[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) =
   "\n\036speechly/config/v1/model.proto\022\022speech"
-  "ly.config.v1\"W\n\tBaseModel\022\014\n\004name\030\001 \001(\t\022"
-  "\r\n\005alias\030\002 \001(\t\022\024\n\014is_available\030\003 \001(\010\022\027\n\017"
-  "is_downloadable\030\004 \001(\010Bs\n\026com.speechly.co"
-  "nfig.v1B\nModelProtoP\001Z\033speechly/config/v"
-  "1;configv1\242\002\003SCX\252\002\022Speechly.Config.V1\312\002\022"
-  "Speechly\\Config\\V1b\006proto3"
+  "ly.config.v1\"A\n\tBaseModel\022\014\n\004name\030\001 \001(\t\022"
+  "\r\n\005alias\030\002 \001(\t\022\027\n\017is_downloadable\030\003 \001(\010B"
+  "s\n\026com.speechly.config.v1B\nModelProtoP\001Z"
+  "\033speechly/config/v1;configv1\242\002\003SCX\252\002\022Spe"
+  "echly.Config.V1\312\002\022Speechly\\Config\\V1b\006pr"
+  "oto3"
   ;
 static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_speechly_2fconfig_2fv1_2fmodel_2eproto_once;
 const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_speechly_2fconfig_2fv1_2fmodel_2eproto = {
-  false, false, 266, descriptor_table_protodef_speechly_2fconfig_2fv1_2fmodel_2eproto, "speechly/config/v1/model.proto", 
+  false, false, 244, descriptor_table_protodef_speechly_2fconfig_2fv1_2fmodel_2eproto, "speechly/config/v1/model.proto", 
   &descriptor_table_speechly_2fconfig_2fv1_2fmodel_2eproto_once, nullptr, 0, 1,
   schemas, file_default_instances, TableStruct_speechly_2fconfig_2fv1_2fmodel_2eproto::offsets,
   file_level_metadata_speechly_2fconfig_2fv1_2fmodel_2eproto, file_level_enum_descriptors_speechly_2fconfig_2fv1_2fmodel_2eproto, file_level_service_descriptors_speechly_2fconfig_2fv1_2fmodel_2eproto,
@@ -114,19 +112,14 @@ BaseModel::BaseModel(const BaseModel& from)
     alias_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, from._internal_alias(), 
       GetArenaForAllocation());
   }
-  ::memcpy(&is_available_, &from.is_available_,
-    static_cast<size_t>(reinterpret_cast<char*>(&is_downloadable_) -
-    reinterpret_cast<char*>(&is_available_)) + sizeof(is_downloadable_));
+  is_downloadable_ = from.is_downloadable_;
   // @@protoc_insertion_point(copy_constructor:speechly.config.v1.BaseModel)
 }
 
 inline void BaseModel::SharedCtor() {
 name_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
 alias_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-::memset(reinterpret_cast<char*>(this) + static_cast<size_t>(
-    reinterpret_cast<char*>(&is_available_) - reinterpret_cast<char*>(this)),
-    0, static_cast<size_t>(reinterpret_cast<char*>(&is_downloadable_) -
-    reinterpret_cast<char*>(&is_available_)) + sizeof(is_downloadable_));
+is_downloadable_ = false;
 }
 
 BaseModel::~BaseModel() {
@@ -160,9 +153,7 @@ void BaseModel::Clear() {
 
   name_.ClearToEmpty();
   alias_.ClearToEmpty();
-  ::memset(&is_available_, 0, static_cast<size_t>(
-      reinterpret_cast<char*>(&is_downloadable_) -
-      reinterpret_cast<char*>(&is_available_)) + sizeof(is_downloadable_));
+  is_downloadable_ = false;
   _internal_metadata_.Clear<::PROTOBUF_NAMESPACE_ID::UnknownFieldSet>();
 }
 
@@ -190,16 +181,9 @@ const char* BaseModel::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID::
           CHK_(ptr);
         } else goto handle_unusual;
         continue;
-      // bool is_available = 3;
+      // bool is_downloadable = 3;
       case 3:
         if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 24)) {
-          is_available_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
-          CHK_(ptr);
-        } else goto handle_unusual;
-        continue;
-      // bool is_downloadable = 4;
-      case 4:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 32)) {
           is_downloadable_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint64(&ptr);
           CHK_(ptr);
         } else goto handle_unusual;
@@ -253,16 +237,10 @@ failure:
         2, this->_internal_alias(), target);
   }
 
-  // bool is_available = 3;
-  if (this->_internal_is_available() != 0) {
-    target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteBoolToArray(3, this->_internal_is_available(), target);
-  }
-
-  // bool is_downloadable = 4;
+  // bool is_downloadable = 3;
   if (this->_internal_is_downloadable() != 0) {
     target = stream->EnsureSpace(target);
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteBoolToArray(4, this->_internal_is_downloadable(), target);
+    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteBoolToArray(3, this->_internal_is_downloadable(), target);
   }
 
   if (PROTOBUF_PREDICT_FALSE(_internal_metadata_.have_unknown_fields())) {
@@ -295,12 +273,7 @@ size_t BaseModel::ByteSizeLong() const {
         this->_internal_alias());
   }
 
-  // bool is_available = 3;
-  if (this->_internal_is_available() != 0) {
-    total_size += 1 + 1;
-  }
-
-  // bool is_downloadable = 4;
+  // bool is_downloadable = 3;
   if (this->_internal_is_downloadable() != 0) {
     total_size += 1 + 1;
   }
@@ -339,9 +312,6 @@ void BaseModel::MergeFrom(const BaseModel& from) {
   if (!from._internal_alias().empty()) {
     _internal_set_alias(from._internal_alias());
   }
-  if (from._internal_is_available() != 0) {
-    _internal_set_is_available(from._internal_is_available());
-  }
   if (from._internal_is_downloadable() != 0) {
     _internal_set_is_downloadable(from._internal_is_downloadable());
   }
@@ -372,12 +342,7 @@ void BaseModel::InternalSwap(BaseModel* other) {
       &alias_, GetArenaForAllocation(),
       &other->alias_, other->GetArenaForAllocation()
   );
-  ::PROTOBUF_NAMESPACE_ID::internal::memswap<
-      PROTOBUF_FIELD_OFFSET(BaseModel, is_downloadable_)
-      + sizeof(BaseModel::is_downloadable_)
-      - PROTOBUF_FIELD_OFFSET(BaseModel, is_available_)>(
-          reinterpret_cast<char*>(&is_available_),
-          reinterpret_cast<char*>(&other->is_available_));
+  swap(is_downloadable_, other->is_downloadable_);
 }
 
 ::PROTOBUF_NAMESPACE_ID::Metadata BaseModel::GetMetadata() const {